Create new cookies module and move classes there
