+++ app/build.gradle
@@ -219,6 +219,9 @@ dependencies {
     implementation project(':autoconsent-impl')
     implementation project(':autoconsent-store')
 
+    implementation project(':cookies-api')
+    implementation project(':cookies-impl')
+
     // Deprecated. TODO: Stop using this artifact.
     implementation "androidx.legacy:legacy-support-v4:_"
     debugImplementation Square.leakCanary.android
