+++ app/src/main/java/com/duckduckgo/app/browser/di/BrowserModule.kt
@@ -28,8 +28,6 @@ import com.duckduckgo.app.browser.addtohome.AddToHomeCapabilityDetector
 import com.duckduckgo.app.browser.addtohome.AddToHomeSystemCapabilityDetector
 import com.duckduckgo.app.browser.certificates.rootstore.TrustedCertificateStore
 import com.duckduckgo.app.browser.cookies.AppThirdPartyCookieManager
-import com.duckduckgo.app.browser.cookies.CookieManagerProvider
-import com.duckduckgo.app.browser.cookies.DefaultCookieManagerProvider
 import com.duckduckgo.app.browser.cookies.ThirdPartyCookieManager
 import com.duckduckgo.app.browser.cookies.db.AuthCookiesAllowedDomainsRepository
 import com.duckduckgo.app.browser.defaultbrowsing.AndroidDefaultBrowserDetector
@@ -54,7 +52,6 @@ import com.duckduckgo.app.browser.urlextraction.UrlExtractingWebViewClient
 import com.duckduckgo.app.browser.useragent.UserAgentProvider
 import com.duckduckgo.app.di.AppCoroutineScope
 import com.duckduckgo.app.fire.*
-import com.duckduckgo.app.fire.fireproofwebsite.data.FireproofWebsiteDao
 import com.duckduckgo.app.fire.fireproofwebsite.data.FireproofWebsiteRepository
 import com.duckduckgo.app.global.AppUrl
 import com.duckduckgo.app.global.DispatcherProvider
@@ -70,7 +67,6 @@ import com.duckduckgo.app.privacy.db.UserAllowListRepository
 import com.duckduckgo.app.referral.AppReferrerDataStore
 import com.duckduckgo.app.settings.db.SettingsDataStore
 import com.duckduckgo.app.statistics.VariantManager
-import com.duckduckgo.app.statistics.pixels.ExceptionPixel
 import com.duckduckgo.app.statistics.pixels.Pixel
 import com.duckduckgo.app.statistics.store.OfflinePixelCountDataStore
 import com.duckduckgo.app.statistics.store.StatisticsDataStore
@@ -82,6 +78,8 @@ import com.duckduckgo.appbuildconfig.api.AppBuildConfig
 import com.duckduckgo.autoconsent.api.Autoconsent
 import com.duckduckgo.autofill.BrowserAutofill
 import com.duckduckgo.autofill.InternalTestUserChecker
+import com.duckduckgo.cookies.api.CookieManagerProvider
+import com.duckduckgo.cookies.api.DuckDuckGoCookieManager
 import com.duckduckgo.di.scopes.AppScope
 import com.duckduckgo.downloads.api.FileDownloader
 import com.duckduckgo.downloads.impl.AndroidFileDownloader
@@ -294,32 +292,8 @@ class BrowserModule {
         )
 
     @Provides
-    fun cookieManager(
-        cookieManagerProvider: CookieManagerProvider,
-        removeCookies: RemoveCookies,
-        dispatcherProvider: DispatcherProvider
-    ): DuckDuckGoCookieManager {
-        return WebViewCookieManager(cookieManagerProvider, AppUrl.Url.COOKIES, removeCookies, dispatcherProvider)
-    }
-
-    @Provides
-    fun removeCookiesStrategy(
-        cookieManagerRemover: CookieManagerRemover,
-        sqlCookieRemover: SQLCookieRemover
-    ): RemoveCookies {
-        return RemoveCookies(cookieManagerRemover, sqlCookieRemover)
-    }
-
-    @Provides
-    fun sqlCookieRemover(
-        @Named("webViewDbLocator") webViewDatabaseLocator: DatabaseLocator,
-        getCookieHostsToPreserve: GetCookieHostsToPreserve,
-        offlinePixelCountDataStore: OfflinePixelCountDataStore,
-        exceptionPixel: ExceptionPixel,
-        dispatcherProvider: DispatcherProvider
-    ): SQLCookieRemover {
-        return SQLCookieRemover(webViewDatabaseLocator, getCookieHostsToPreserve, offlinePixelCountDataStore, exceptionPixel, dispatcherProvider)
-    }
+    @Named("cookiesUrl")
+    fun cookiesUrl(): String = AppUrl.Url.COOKIES
 
     @Provides
     @Named("webViewDbLocator")
@@ -331,22 +305,6 @@ class BrowserModule {
 
     @Provides
     fun databaseCleanerHelper(): DatabaseCleaner = DatabaseCleanerHelper()
-
-    @Provides
-    fun getCookieHostsToPreserve(fireproofWebsiteDao: FireproofWebsiteDao): GetCookieHostsToPreserve = GetCookieHostsToPreserve(fireproofWebsiteDao)
-
-    @Provides
-    fun cookieManagerRemover(
-        cookieManagerProvider: CookieManagerProvider
-    ): CookieManagerRemover {
-        return CookieManagerRemover(cookieManagerProvider)
-    }
-
-    @SingleInstanceIn(AppScope::class)
-    @Provides
-    fun webViewCookieManagerProvider(): CookieManagerProvider {
-        return DefaultCookieManagerProvider()
-    }
 
     @SingleInstanceIn(AppScope::class)
     @Provides
