--- app/src/main/java/com/duckduckgo/app/fire/GetCookieHostsToPreserve.kt
@@ -1,40 +1,0 @@
-/*
- * Copyright (c) 2020 DuckDuckGo
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package com.duckduckgo.app.fire
-
-import com.duckduckgo.app.fire.fireproofwebsite.data.FireproofWebsiteDao
-
-class GetCookieHostsToPreserve(private val fireproofWebsiteDao: FireproofWebsiteDao) {
-    operator fun invoke(): List<String> {
-        val fireproofWebsites = fireproofWebsiteDao.fireproofWebsitesSync()
-        return fireproofWebsites.flatMap { entity ->
-            val acceptedHosts = mutableSetOf<String>()
-            val host = entity.domain
-            acceptedHosts.add(host)
-            host.split(".")
-                .foldRight(
-                    "",
-                    { next, acc ->
-                        val acceptedHost = ".$next$acc"
-                        acceptedHosts.add(acceptedHost)
-                        acceptedHost
-                    }
-                )
-            acceptedHosts
-        }.distinct()
-    }
-}
