+++ app/src/main/java/com/duckduckgo/app/fire/fireproofwebsite/data/FireproofWebsiteRepository.kt
@@ -19,17 +19,40 @@ package com.duckduckgo.app.fire.fireproofwebsite.data
 import android.net.Uri
 import androidx.lifecycle.LiveData
 import com.duckduckgo.app.browser.favicon.FaviconManager
+import com.duckduckgo.app.fire.FireproofRepository
 import com.duckduckgo.app.global.DispatcherProvider
 import com.duckduckgo.app.global.UriString
+import com.duckduckgo.di.scopes.AppScope
+import com.squareup.anvil.annotations.ContributesBinding
 import dagger.Lazy
 import kotlinx.coroutines.withContext
 import javax.inject.Inject
 
+@ContributesBinding(AppScope::class)
 class FireproofWebsiteRepository @Inject constructor(
     private val fireproofWebsiteDao: FireproofWebsiteDao,
     private val dispatchers: DispatcherProvider,
     private val faviconManager: Lazy<FaviconManager>
-) {
+) : FireproofRepository {
+
+    override fun fireproofWebsites(): List<String> {
+        val fireproofWebsites = fireproofWebsiteDao.fireproofWebsitesSync()
+        return fireproofWebsites.flatMap { entity ->
+            val acceptedHosts = mutableSetOf<String>()
+            val host = entity.domain
+            acceptedHosts.add(host)
+            host.split(".")
+                .foldRight(
+                    ""
+                ) { next, acc ->
+                    val acceptedHost = ".$next$acc"
+                    acceptedHosts.add(acceptedHost)
+                    acceptedHost
+                }
+            acceptedHosts
+        }.distinct()
+    }
+
     suspend fun fireproofWebsite(domain: String): FireproofWebsiteEntity? {
         if (!UriString.isValidDomain(domain)) return null
