+++ app/src/main/java/com/duckduckgo/app/pixels/AppPixelName.kt
@@ -192,9 +192,6 @@ enum class AppPixelName(override val pixelName: String) : Pixel.PixelName {
     MENU_ACTION_APP_LINKS_OPEN_PRESSED("m_nav_app_links_open_menu_item_pressed"),
     MENU_ACTION_DOWNLOADS_PRESSED("m_nav_downloads_menu_item_pressed"),
 
-    COOKIE_DATABASE_EXCEPTION_OPEN_ERROR("m_cdb_e_oe"),
-    COOKIE_DATABASE_EXCEPTION_DELETE_ERROR("m_cdb_e_de"),
-
     FIREPROOF_WEBSITE_ADDED("m_fw_a"),
     FIREPROOF_WEBSITE_REMOVE("m_fw_r"),
     FIREPROOF_LOGIN_DIALOG_SHOWN("m_fw_ld_s"),
